proshield-reports/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.1
Flask-Login>=0.6.3
Flask-WTF>=1.2.1
Werkzeug>=3.0.1
Pillow>=10.0.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
openpyxl>=3.1.2
